src/foundry-agent/requirements.txt
azure-identity>=1.18,<2.0
openai>=1.52,<2.0
fastapi>=0.115,<1.0
uvicorn>=0.32,<1.0
